+++ Assets/Scripts/Weapons/Gun.cs
@@ -7,7 +7,7 @@ public class Gun : MonoBehaviour
     // Start is called before the first frame update
 
 
-    private int _fireRate; // shots per second
+    private float _fireRate; // shots per second
     private int _bulletType; // 0 = normal - uses dictionary
     private int _mods; // 0 = no mods TODO: could look into using bitwise to get unique combinations
     private int _currentClip;
